# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.PaginationInfo:
    properties:
      is_pager:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  http.R:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  http.RScript:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.ScriptResp'
      msg:
        type: string
    type: object
  http.RScriptList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.ScriptListResp'
      msg:
        type: string
    type: object
  http.RTenant:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.TenantResp'
      msg:
        type: string
    type: object
  http.RTenantList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.TenantListResp'
      msg:
        type: string
    type: object
  http.RTenantUserList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.TenantUserListResp'
      msg:
        type: string
    type: object
  http.ScriptListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.ScriptResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.ScriptResp:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      tenant_id:
        type: integer
      type:
        type: string
    type: object
  http.TenantListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.TenantResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.TenantResp:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  http.TenantUserListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.TenantUserResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.TenantUserResp:
    properties:
      joined_at:
        type: string
      role:
        type: string
      tenant_id:
        type: integer
      user:
        $ref: '#/definitions/http.UserResp'
      user_id:
        type: integer
    type: object
  http.UserResp:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_superadmin:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: 企业级 SQL to API 智能网关平台管理接口
  title: OwlApi Control Plane
  version: "1.0"
paths:
...
  /v1/my/tenants:
    get:
      operationId: myTenants
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenantList'
      security:
      - BearerAuth: []
      summary: 获取当前用户的租户列表
      tags:
      - tenant
  /v1/tenants:
    get:
      operationId: listAllTenants
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenantList'
      security:
      - BearerAuth: []
      summary: 获取所有租户 (SuperAdmin)
      tags:
      - tenant
    post:
      consumes:
      - application/json
      operationId: createTenant
      parameters:
      - description: 租户信息
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
            plan:
              type: string
            slug:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 创建租户 (SuperAdmin)
      tags:
      - tenant
  /v1/tenants/{slug}:
    delete:
      operationId: deleteTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除租户 (SuperAdmin)
      tags:
      - tenant
    get:
      operationId: getTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 获取租户详情
      tags:
      - tenant
    put:
      consumes:
      - application/json
      operationId: updateTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
            plan:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 更新租户 (SuperAdmin)
      tags:
      - tenant
...
  /v1/tenants/{slug}/scripts:
    get:
      operationId: listScripts
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RScriptList'
      security:
      - BearerAuth: []
      summary: 获取脚本列表
      tags:
      - script
    post:
      consumes:
      - application/json
      operationId: createScript
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 脚本信息
        in: body
        name: body
        required: true
        schema:
          properties:
            code:
              type: string
            description:
              type: string
            name:
              type: string
            type:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RScript'
      security:
      - BearerAuth: []
      summary: 创建脚本
      tags:
      - script
  /v1/tenants/{slug}/scripts/{scriptId}:
    delete:
      operationId: deleteScript
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: scriptId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除脚本
      tags:
      - script
    put:
      consumes:
      - application/json
      operationId: updateScript
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: scriptId
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        schema:
          properties:
            code:
              type: string
            description:
              type: string
            name:
              type: string
            type:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RScript'
      security:
      - BearerAuth: []
      summary: 更新脚本
      tags:
      - script
  /v1/tenants/{slug}/users:
    get:
      operationId: listUsers
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenantUserList'
      security:
      - BearerAuth: []
      summary: 获取租户成员列表
      tags:
      - tenant-user
    post:
      consumes:
      - application/json
      operationId: addUser
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 成员信息
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 添加租户成员
      tags:
      - tenant-user
  /v1/tenants/{slug}/users/{userId}:
    delete:
      operationId: removeUser
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 移除租户成员
      tags:
      - tenant-user
  /v1/tenants/{slug}/users/{userId}/role:
    put:
      consumes:
      - application/json
      operationId: updateUserRole
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 角色
        in: body
        name: body
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 更新成员角色
      tags:
      - tenant-user
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"